+++ src/Manager/Pdf/VehicleMaintenancePdfManager.php
@@ -52,27 +52,29 @@ class VehicleMaintenancePdfManager
      */
     private function buildVehicleMaintenancePdf($vehicleMaintenances, \TCPDF $pdf): \TCPDF
     {
-        //        $this->addStartPage($pdf);
+        $this->addStartPage($pdf);
         $vehiclesFromVehicleMaintenances = [];
         /** @var VehicleMaintenance $vehicleMaintenance */
         foreach ($vehicleMaintenances as $vehicleMaintenance) {
             $vehiclesFromVehicleMaintenances[$vehicleMaintenance->getVehicle()->getId()][] = $vehicleMaintenance;
         }
 
+        $pdf->setXY(30, 40);
+
         foreach ($vehiclesFromVehicleMaintenances as $vehicleId => $vehicleMaintenancesFromVehicle) {
             /** @var Vehicle $vehicle */
             $vehicle = $this->rm->getVehicleRepository()->find($vehicleId);
-            //            if ($pdf->getY() > 210) {
-            $this->addStartPage($pdf);
-            //            }
+            if ($pdf->getY() > 220) {
+                $this->addStartPage($pdf);
+                $pdf->setXY(30, 40);
+            }
             // Header
-            $pdf->setXY(30, 40);
             $pdf->setX(ConstantsEnum::PDF_PAGE_A4_MARGIN_LEFT);
             $this->pdfEngineService->setStyleSize('b', 10);
             $pdf->Cell(110, ConstantsEnum::PDF_CELL_HEIGHT,
                 'Vehículo - '.$vehicle->getName(),
                 true, 0, 'L', true);
-            $pdf->Cell(75, ConstantsEnum::PDF_CELL_HEIGHT,
+            $pdf->Cell(80, ConstantsEnum::PDF_CELL_HEIGHT,
                 'Matrícula - '.$vehicle->getVehicleRegistrationNumber(),
                 true, 0, 'L', true);
             $pdf->Ln();
@@ -85,14 +87,17 @@ class VehicleMaintenancePdfManager
             $pdf->Cell(50, ConstantsEnum::PDF_CELL_HEIGHT,
                 'Descripción',
                 true, 0, 'L', true);
-            $pdf->Cell(25, ConstantsEnum::PDF_CELL_HEIGHT,
-                'Km restantes',
+            $pdf->Cell(20, ConstantsEnum::PDF_CELL_HEIGHT,
+                'Última rev.',
                 true, 0, 'L', true);
-            $pdf->Cell(25, ConstantsEnum::PDF_CELL_HEIGHT,
-                'Horas restantes',
+            $pdf->Cell(20, ConstantsEnum::PDF_CELL_HEIGHT,
+                'Km rest.',
                 true, 0, 'L', true);
-            $pdf->Cell(25, ConstantsEnum::PDF_CELL_HEIGHT,
-                'Fecha efectuada',
+            $pdf->Cell(20, ConstantsEnum::PDF_CELL_HEIGHT,
+                'Horas rest.',
+                true, 0, 'L', true);
+            $pdf->Cell(20, ConstantsEnum::PDF_CELL_HEIGHT,
+                'Fecha rev.',
                 true, 0, 'L', true);
             $pdf->Ln();
             // info
@@ -106,17 +111,21 @@ class VehicleMaintenancePdfManager
                 $pdf->Cell(50, ConstantsEnum::PDF_CELL_HEIGHT,
                     $vehicleMaintenanceFromVehicle->getDescription(),
                     true, 0, 'L', false, '', 1);
-                $pdf->Cell(25, ConstantsEnum::PDF_CELL_HEIGHT,
+                $pdf->Cell(20, ConstantsEnum::PDF_CELL_HEIGHT,
+                    $vehicleMaintenanceFromVehicle->getDate()->format('d/m/Y'),
+                    true, 0, 'L', false);
+                $pdf->Cell(20, ConstantsEnum::PDF_CELL_HEIGHT,
                     $this->vehicleMaintenanceManager->remainingKm($vehicleMaintenanceFromVehicle),
                     true, 0, 'L', false);
-                $pdf->Cell(25, ConstantsEnum::PDF_CELL_HEIGHT,
+                $pdf->Cell(20, ConstantsEnum::PDF_CELL_HEIGHT,
                     $this->vehicleMaintenanceManager->remainingHours($vehicleMaintenanceFromVehicle),
                     true, 0, 'L', false);
-                $pdf->Cell(25, ConstantsEnum::PDF_CELL_HEIGHT,
+                $pdf->Cell(20, ConstantsEnum::PDF_CELL_HEIGHT,
                     '',
                     true, 0, 'L', false);
                 $pdf->Ln();
             }
+            $pdf->Ln();
         }
 
         return $pdf;